Issue #81 refactor: id 타입 수정
Int -> Long

+++ app/src/main/java/com/drunkenboys/calendarun/ui/savecalendar/SaveCalendarViewModel.kt
@@ -75,7 +75,7 @@ class SaveCalendarViewModel @Inject constructor(
         _checkPointItemList.value = newList
     }
 
-    fun fetchCalendar(id: Int) {
+    fun fetchCalendar(id: Long) {
         viewModelScope.launch {
             val selectedCalendar = calendarLocalDataSource.fetchCalendar(id)
